SkillSelectionAdapter created
