+++ app/src/main/java/org/oppia/app/topic/train/SkillSelectionAdapter.kt
@@ -1,0 +1,47 @@
+package org.oppia.app.topic.train;
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.databinding.library.baseAdapters.BR
+import androidx.recyclerview.widget.RecyclerView
+import kotlinx.android.synthetic.main.topic_train_skill_view.view.*
+import org.oppia.app.R
+import org.oppia.app.databinding.TopicTrainSkillViewBinding
+
+/** Adapter to bind skills to [RecyclerView] inside [TopicTrainFragment]. **/
+class SkillSelectionAdapter(private val context: Context, private val itemList: List<String>) :
+  RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+
+  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    val inflater = LayoutInflater.from(parent.context)
+    val binding =
+      DataBindingUtil.inflate<TopicTrainSkillViewBinding>(
+        inflater,
+        R.layout.topic_train_skill_view,
+        parent,
+        false
+      )
+    return SkillViewHolder(binding)
+  }
+
+  override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+    (holder as SkillViewHolder).bind(itemList[position], position)
+  }
+
+  override fun getItemCount(): Int {
+    return itemList.size
+  }
+
+  private inner class SkillViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
+    internal fun bind(rawString: String?, position: Int) {
+      binding.setVariable(BR.skill, rawString)
+      binding.root.skill_check_box.setOnClickListener {
+        Toast.makeText(context, "" + binding.root.skill_check_box.text, Toast.LENGTH_LONG).show()
+      }
+    }
+  }
+}
