+++ app/src/main/res/layout/topic_train_fragment.xml
@@ -45,7 +45,6 @@
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
           android:layout_marginTop="16dp"
-          app:data="@{viewModel.skillListLiveData}"
           app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"/>
         <Button
           android:id="@+id/topic_train_start_button"
